Moved notion import business logic inside lib/notion folder
